app/pyproject.toml
[tool.poetry]
name = "ai-evaluation-bot"
version = "0.1.0"
description = "A Slack bot built with FastAPI"
authors = ["David Podolskyi <podolskydave@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.0"
slack-sdk = "^3.27.0"
python-multipart = "^0.0.6"
docling = "^2.15.1"
haystack-ai = "^2.9.0"
requests = "^2.31.0"
openai = "^1.60.2"
anthropic = "^0.45.2"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
pytest = "^8.0.0"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"